# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PetMon API
  description: Pet monitoring and nutrition tracking API
  version: 1.0.0
  contact:
    name: PetMon Team
  license:
    name: MIT
servers:
  - url: /api/v1
    description: Main API server
security:
  - bearerAuth: []
paths:
...
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user identity
      description: Returns information about the authenticated user making the request
      responses:
        '200':
          description: Current user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /days/{date}:
    get:
      tags:
        - Days
      summary: Get day summary
      x-required-scope: api_read
      description: Retrieve nutrition records and summary for a specific date. Optionally filter by pet.
      parameters:
        - name: date
          in: path
          required: true
          description: Date in YYYY-MM-DD format
          schema:
            type: string
            format: date
        - name: pet_id
          in: query
          required: false
          description: Filter by pet UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Day summary with nutrition records and totals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NutritionDaySummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /days/{date}/note:
    patch:
      tags:
        - Days
      summary: Update day note
      x-required-scope: api_write
      description: Update or create a note for a specific day. Optionally scoped to a pet.
      parameters:
        - name: date
          in: path
          required: true
          description: Date in YYYY-MM-DD format
          schema:
            type: string
            format: date
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - note
              properties:
                note:
                  type: string
                  description: Note text
                pet_id:
                  type: string
                  format: uuid
                  description: Optional pet UUID to scope the note
      responses:
        '200':
          description: Note updated successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /elimination/analytics/daily-summaries:
    get:
      tags:
        - Elimination Analytics
      summary: Get daily elimination summaries
      x-required-scope: api_read
      parameters:
        - name: date_from
          in: query
          required: true
          schema: { type: string, format: date }
        - name: date_to
          in: query
          required: true
          schema: { type: string, format: date }
        - name: pet_id
          in: query
          required: false
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Daily summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EliminationDailySummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /elimination/analytics/range-summary:
    get:
      tags:
        - Elimination Analytics
      summary: Get elimination range summary with percentiles
      x-required-scope: api_read
      parameters:
        - name: date_from
          in: query
          required: true
          schema: { type: string, format: date }
        - name: date_to
          in: query
          required: true
          schema: { type: string, format: date }
        - name: pet_id
          in: query
          required: false
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Range summary with avg/p50/p90/p99 per day
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EliminationRangeSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /nutrition/analytics/range-summary:
    get:
      tags:
        - Nutrition Analytics
      summary: Get range summary
      x-required-scope: api_read
      description: Get aggregated nutrition summary for a date range with category averages
      parameters:
        - name: pet_id
          in: query
          required: false
          description: Filter by pet UUID
          schema:
            type: string
            format: uuid
        - name: date_from
          in: query
          required: true
          description: Start date (YYYY-MM-DD, inclusive)
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          required: true
          description: End date (YYYY-MM-DD, inclusive)
          schema:
            type: string
            format: date
        - name: category
          in: query
          required: false
          description: Filter by nutrition category
          schema:
            $ref: '#/components/schemas/NutritionCategory'
      responses:
        '200':
          description: Range summary with aggregated data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NutritionRangeSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Bearer authentication with either:
        - OIDC JWT token (format: standard JWT) — full access, no scope restrictions
        - API token with `pm_api_` prefix — access controlled by the token's scope set

        **API token scopes:**
        | Scope | Access granted |
        |---|---|
        | `all` | Full access — equivalent to OIDC |
        | `api_read` | All GET endpoints |
        | `api_write` | All POST / PATCH / DELETE endpoints |
        | `mcp` | MCP tool endpoint (`/mcp`) |

        A token may hold multiple scopes (e.g. `["api_read", "mcp"]`).
        The scope required by each endpoint is shown in the `x-required-scope` field.
  responses:
    BadRequest:
      description: Invalid request parameters or body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Token does not have the required scope for this endpoint
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required:
        - error
...
    MeResponse:
      type: object
      properties:
        subject:
          type: string
          description: User subject identifier
        email:
          type: string
          nullable: true
          description: User email address
        name:
          type: string
          nullable: true
          description: User full name
        display_name:
          type: string
          description: Display name (name or email or subject)
        kind:
          type: string
          enum:
            - oidc
            - api_token
            - dev
          description: Authentication method used
      required:
        - subject
        - display_name
        - kind
...
    NutritionCategory:
      type: string
      enum:
        - wet_food
        - dry_food
        - water
        - liquids
      description: Category of nutrition intake
    NutritionRecord:
      type: object
      properties:
        id:
          type: string
          description: Record ID
        pet_id:
          type: string
          format: uuid
          description: Pet UUID
        occurred_at:
          type: string
          format: date-time
          description: When the intake occurred (RFC3339)
        local_date:
          type: string
          format: date
          description: Local date of the occurrence
        category:
          $ref: '#/components/schemas/NutritionCategory'
        amount:
          type: number
          format: double
          description: Amount consumed
        unit:
          type: string
          nullable: true
          description: Unit of measurement (ml, g, etc.)
        source_type:
          type: string
          description: Source of the record (e.g., manual, telegram)
        created_at:
          type: string
          format: date-time
          description: Record creation timestamp (RFC3339)
        updated_at:
          type: string
          format: date-time
          description: Record last update timestamp (RFC3339)
      required:
        - id
        - pet_id
        - occurred_at
        - local_date
        - category
        - amount
        - source_type
        - created_at
        - updated_at
...
    NutritionDailyTotal:
      type: object
      properties:
        local_date:
          type: string
          format: date
          description: Date of the totals
        pet_id:
          type: string
          format: uuid
          description: Pet UUID
        category:
          type: string
          description: Nutrition category
        total_amount:
          type: number
          format: double
          description: Total amount for the day
        record_count:
          type: integer
          format: int64
          description: Number of records contributing to this total
      required:
        - local_date
        - pet_id
        - category
        - total_amount
        - record_count
    NutritionDaySummary:
      type: object
      properties:
        local_date:
          type: string
          format: date
          description: Date of the summary
        pet_id:
          type: string
          format: uuid
          nullable: true
          description: Pet UUID (if filtered)
        records:
          type: array
          items:
            $ref: '#/components/schemas/NutritionRecord'
          description: Nutrition records for the day
        totals_by_category:
          type: object
          additionalProperties:
            type: number
            format: double
          description: Total amounts by category
        note:
          type: string
          nullable: true
          description: Optional day note
      required:
        - local_date
        - records
        - totals_by_category
...
    NutritionRangeSummary:
      type: object
      properties:
        date_from:
          type: string
          format: date
          description: Start date of the range
        date_to:
          type: string
          format: date
          description: End date of the range
        pet_id:
          type: string
          format: uuid
          nullable: true
          description: Pet UUID (if filtered)
        daily_totals:
          type: array
          items:
            $ref: '#/components/schemas/NutritionDailyTotal'
          description: Daily totals for the date range
        category_averages:
          type: object
          additionalProperties:
            type: number
            format: double
          description: Average daily amounts by category
      required:
        - date_from
        - date_to
        - daily_totals
        - category_averages
...
    EliminationDailySummary:
      type: object
      properties:
        local_date:
          type: string
          format: date
        pet_id:
          type: string
          format: uuid
          nullable: true
        total_count:
          type: integer
          format: int64
        urination_count:
          type: integer
          format: int64
        defecation_count:
          type: integer
          format: int64
        vomit_count:
          type: integer
          format: int64
        general_count:
          type: integer
          format: int64
        has_vomit:
          type: boolean
        avg_duration_seconds:
          type: number
          format: double
          nullable: true
          description: Average session duration in seconds for records that have duration logged; null if none.
      required:
        - local_date
        - total_count
        - urination_count
        - defecation_count
        - vomit_count
        - general_count
        - has_vomit
    EliminationRangeSummary:
      type: object
      properties:
        date_from:
          type: string
          format: date
        date_to:
          type: string
          format: date
        pet_id:
          type: string
          format: uuid
          nullable: true
        daily_summaries:
          type: array
          items:
            $ref: '#/components/schemas/EliminationDailySummary'
        type_totals:
          type: object
          additionalProperties:
            type: integer
            format: int64
        avg_per_day:
          type: number
          format: double
        p50_per_day:
          type: number
          format: double
        p90_per_day:
          type: number
          format: double
        p99_per_day:
          type: number
          format: double
      required:
        - date_from
        - date_to
        - daily_summaries
        - type_totals
        - avg_per_day
        - p50_per_day
        - p90_per_day
        - p99_per_day
...